cloud_functions/gemini-pro-news/requirements.txt
functions-framework==3.*
google-generativeai
gnews
requests